Updated Table Generator
Made it so that if you did not have a table made already, you can select option 2 in the menu to create a table.

Fixed various bugs related to querying the data. A few sections that are in the DB are very messy, and don't have the correct structure. (Might make a fix for that in the future, so that you can still query the database, but it will be inserted into your own as consistent data)

Added example_structures of each table that's being queried. This is so that you can understand how you should set up your own projects, and handling the data.

Completely refactored how we're inserting the data into our own database. This has better error handling, and retry code for timeouts are more accurate. This also includes getting the `_id` in place of `guid` where guid doesn't exist, and _id does (mainly Takyn's workaround for ashes not having consistent data). As well as _slug (Which is thankfully added by Takyn) for mobs, and hunting-creatures. (The _slug key will be used as the guid in those instances)

Updated the .env creator text so that you must input your database information if you want to create the table from script. This could be handy, as the tool could later be updated to handle different db's, or raw querying from CLI.